+++ packages/expo-router/src/layouts/stack-utils/index.tsx
@@ -1,9 +1,29 @@
 import { StackHeaderBackButton, type StackHeaderBackButtonProps } from './StackHeaderBackButton';
+import { StackHeaderButton, type StackHeaderButtonProps } from './StackHeaderButton';
 import { StackHeaderComponent, type StackHeaderProps } from './StackHeaderComponent';
-import { StackHeaderLeft, type StackHeaderLeftProps } from './StackHeaderLeft';
-import { StackHeaderRight, type StackHeaderRightProps } from './StackHeaderRight';
+import { StackHeaderItem, type StackHeaderItemProps } from './StackHeaderItem';
+import {
+  StackHeaderLeft,
+  type StackHeaderLeftProps,
+  StackHeaderRight,
+  type StackHeaderRightProps,
+} from './StackHeaderLeftRight';
+import {
+  StackHeaderMenu,
+  StackHeaderMenuAction,
+  type StackHeaderMenuActionProps,
+  type StackHeaderMenuProps,
+} from './StackHeaderMenu';
 import { StackHeaderSearchBar, type StackHeaderSearchBarProps } from './StackHeaderSearchBar';
 import { StackHeaderTitle, type StackHeaderTitleProps } from './StackHeaderTitle';
+import {
+  StackHeaderBadge,
+  StackHeaderIcon,
+  StackHeaderLabel,
+  type StackHeaderBadgeProps,
+  type StackHeaderIconProps,
+  type StackHeaderLabelProps,
+} from './common-primitives';
 
 export const StackHeader = Object.assign(StackHeaderComponent, {
   Left: StackHeaderLeft,
@@ -11,6 +31,13 @@ export const StackHeader = Object.assign(StackHeaderComponent, {
   BackButton: StackHeaderBackButton,
   Title: StackHeaderTitle,
   SearchBar: StackHeaderSearchBar,
+  Button: StackHeaderButton,
+  Badge: StackHeaderBadge,
+  Label: StackHeaderLabel,
+  Icon: StackHeaderIcon,
+  Menu: StackHeaderMenu,
+  MenuAction: StackHeaderMenuAction,
+  Item: StackHeaderItem,
 });
 
 export {
@@ -26,6 +53,21 @@ export {
   type StackHeaderSearchBarProps,
   StackHeaderTitle,
   type StackHeaderTitleProps,
+  StackHeaderButton,
+  type StackHeaderButtonProps,
+  StackHeaderBadge,
+  type StackHeaderBadgeProps,
+  StackHeaderLabel,
+  type StackHeaderLabelProps,
+  StackHeaderIcon,
+  type StackHeaderIconProps,
+  StackHeaderMenu,
+  type StackHeaderMenuProps,
+  StackHeaderMenuAction,
+  type StackHeaderMenuActionProps,
+  StackHeaderItem,
+  type StackHeaderItemProps,
 };
+
 export { StackScreen, appendScreenStackPropsToOptions, type StackScreenProps } from './StackScreen';
 export { isChildOfType } from './utils';